♻️ Refactor :  useLogout -> useAuth 리팩토링
