+++ src/domains/user/queries/useLoginMutation.ts
@@ -1,15 +1,11 @@
 import { useRouter } from 'next/navigation';
-import { useSetRecoilState } from 'recoil';
 import { useMutation } from '@tanstack/react-query';
-import { isLoggedinState } from '@/shared/atoms/login';
 import fetchExtend from '@/shared/utils/api';
 import PATH from '@/shared/constants/path';
-import { setCookie } from '@/shared/actions/cookie';
 import { ResultResponse } from '@/shared/types/response';
 import { throwApiError } from '@/shared/utils/error';
-import { TOKEN } from '@/shared/constants/token';
-import { getExpFromToken } from '@/domains/user/utils/jwt';
 import useToastNewVer from '@/shared/hooks/useToastNewVer';
+import useAuth from '@/shared/hooks/useAuth';
 
 export interface LoginResponse {
   accessToken: string;
@@ -18,8 +14,8 @@ export interface LoginResponse {
 
 const useLoginMutation = () => {
   const { openToast } = useToastNewVer();
-  const setLogin = useSetRecoilState(isLoggedinState);
   const { replace } = useRouter();
+  const { login } = useAuth();
 
   const mutationFn = async (accessToken: string) => {
     const res = await fetchExtend.get(`/oauth/login/kakao?token=${accessToken}`);
@@ -31,24 +27,8 @@ const useLoginMutation = () => {
   };
 
   const onSuccess = async ({ accessToken, refreshToken }: LoginResponse) => {
-    const accessTokenExp = getExpFromToken(accessToken);
-    const refreshTokenExp = getExpFromToken(refreshToken);
-
-    await Promise.all([
-      setCookie({
-        name: TOKEN.accessToken,
-        value: accessToken,
-        expires: accessTokenExp
-      }),
-      setCookie({
-        name: TOKEN.refreshToken,
-        value: refreshToken,
-        expires: refreshTokenExp
-      })
-    ]);
+    await login({ accessToken, refreshToken });
     openToast({ message: '로그인 되었어요.' });
-
-    setLogin(true);
     replace(PATH.home);
   };
